+++ examples/miscellaneous/redirect_output.rs
@@ -1,6 +1,3 @@
-// [dependencies]
-// kdam = { version = "0.3.0", features = ["writer"] }
-
 use kdam::{tqdm, BarExt};
 use std::fs::File;
 use std::io::Write;
@@ -12,7 +9,7 @@ fn main() {
 
     for _ in 0..100 {
         std::thread::sleep(std::time::Duration::from_secs_f32(0.01));
-        pb.update_writer(1, &mut f);
+        pb.write_to(&mut f, Some(1));
     }
 
     f.write_all("Finished Writing Logs\n".as_bytes()).unwrap();